backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
dropbox
pandas
